Updated Solution - GfG to GitHub

+++ Difficulty: Medium/Kadane's Algorithm/kadanes-algorithm.cpp
@@ -9,10 +9,10 @@ using namespace std;
 class Solution {
   public:
     // Function to find the sum of contiguous subarray with maximum sum.
-    long long maxSubarraySum(vector<int> &arr) {
+    int maxSubarraySum(vector<int> &arr) {
         // code here...
-        long long sum=0;
-        long long maxi=arr[0];
+        int sum=0;
+        int maxi=arr[0];
         for(int i=0;i<arr.size();i++){
             sum+=arr[i];
             maxi=max(maxi,sum);
